module/pyproject.toml
[tool.poetry]
name = "keyphrase_extractors"
version = "0.1.0"
description = "Keyphrase extraction tools in Japanese"
authors = ["flatton <flatton211@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"
packages = [{ include = "keyphrase_extractors", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.8.2"
neologdn = "^0.5.3"
ja-ginza = "^5.2.0"
pke = { git = "https://github.com/boudinfl/pke.git" }
keybert = "^0.8.5"
sentencepiece = "^0.2.0"
fugashi = "^1.4.0"
unidic-lite = "^1.0.8"
scipy-stubs = "^1.15.1.0"
langrila = { extras = ["openai"], version = "^1.2.1" }
python-dotenv = "^1.0.1"
rapidfuzz = "^3.11.0"
pandas-stubs = "^2.2.3.241126"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
